refactor Generator class works directly with dataset.
Trained model successfully.
